backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
azure-cosmos>=4.5.0
azure-identity>=1.15.0
azure-core>=1.30.0
colorama
gunicorn==21.2.0
python-jose
requests>=2.31.0